A thread must be joined to release it's resources
The started flag will be set to 0 when the thread proc exists.
If lw_join is called after this then the pthread_join wouldn't
be called if we checked the started flag. This would cause a resource
leak since join is required to be called to release the thread
resources since lw_thread is using joinable pthreads